Inline recursive functions w/non-recursive wrapper.
Inlining is often really useful for performance - it's cheaper to just
execute some code than it is to call a function that executes said code.
But GHC can't inline recursive functions.

It's more efficient to write recursive functions using a kind of
worker/wrapper pattern.  The 'loop' function is the recursive worker
that's wrapped up, and the wrapped-up function can be inlined using an
{-# INLINE #-} pragma.