Add profiling output.
GHC has awesome profiling capabilities to figure out exactly where time
is being spent, or which functions are responsibly for what amount of
allocation.

Here you can see that 'acquireSeedSystem' is taking up 81% of the time
of the program, whereas 'withSystemRandom' is responsible for 42% of the
allocation.

We can improve the situation here by only acquiring the system seed
*once*, and only using withSystemRandom to initialize the generator
once.