Add cabal and stack files.
The cabal file specifies your project.  I've plunked 'NestedSampling-hs'
into a library and 'ns-example' into an executable.  You can see the
dependencies, exposed modules, etc. listed in the appropriate places.

The stack file is a prerequisite for using stack (www.haskellstack.org).
It specifies a resolver for a human-curated list of dependencies that
are guaranteed to work together well.

To build things:

$ stack build

To install 'ns-example' into a local bin directory:

$ stack install NestedSampling-hs:ns-example

To compile for profiling:

$ stack install NestedSampling-hs:ns-example \
    --executable-profiling --library-profiling \
    --ghc-options="-fprof-auto -rtsopts -O2"