Use MaybeT to defer error handling.
Here I've used 'MaybeT' to handle error conditions in
'nestedSamplingIterations'.  A crash course in 'Maybe' and monad
transformers is below.

'Maybe' is great for handling missing value cases, like when you want to
grab something from a list but the list could be empty.  It provides a
way to replace the 'null' of C/C++/Java/Go/etc. with a more manageable
alternative.

    safeHead :: [a] -> Maybe a
    safeHead list = case list of
      []    -> Nothing
      (x:_) -> Just x

    safeLast :: [a] -> Maybe a
    safeLast list = case list of
      []     -> Nothing
      (x:[]) -> Just x
      xs     -> safeLast xs

Maybe is a monad like IO.  You can use do-notation to program as if
whatever you're doing doesn't involve 'Nothing' and 'Just' and all that:

    bounds :: [a] -> Maybe (a, a)
    bounds list = do
      h <- safeHead list
      l <- safeLast list
      return (h, l)

'bounds' will return 'Nothing' if either of 'safeHead' or 'safeLast' fail,
and will otherwise return 'Just (h, l)' for h and l the first and last
elements respectively.

You can use *monad transformers* to layer monads together, giving you a
bunch of onion-like layers of functionality.  It's a super useful
pattern.

For a contrived example, say you want to print information about what's
happening to stdout while you go about your business in 'bounds'.  The
following won't work:

    loggingBounds :: [a] -> ???
    loggingBounds list = do
      h <- safeHead list
      putStrLn $ "list had a first element"
      l <- safeLast list
      putStrLn $ "list had a last element"
      return (h, l)

It won't work because you're trying to use do-notation to deal with both
the Maybe and IO monads simultaneously.  What type would that even
correspond to?

The 'MaybeT' monad transformer layers functionality on top of 'IO' so
that we can do that kind of thing, though:

    loggingBounds :: [a] -> MaybeT IO (a, a)
    loggingBounds list = do
      h <- hoistMaybe (safeHead list)
      lift . putStrLn $ "list had a first element"
      l <- hoistMaybe (safeLast list)
      lift . putStrLn $ "list had a last element"
      return (h, l)

The type 'MaybeT IO (a, a)' is a monad with two layers; the outer layer
is 'MaybeT' and the inner layer is 'IO'.  In this example, to access the
'Maybe layer' we use 'hoistMaybe', and to access the 'IO layer' we use
'lift'.

When it comes time to handle the error, we can just do it like so:

    main :: IO ()
    main = do
      let xs = [1, 2, 3]
      result <- runMaybeT (loggingBounds xs)
      case result of
        Nothing     -> putStrLn "empty list!"
        Just bounds -> putStrLn "everything went swimmingly"