Allow model-specific perturbation functions.
Resolves #9.

Here I've done the same kind of pattern and added a 'perturber'
argument to functions that require it.  The user only needs to pass the
thing once, when initializing the Sampler.

This setup will probably need to be tweaked depending on how general you
want to allow the perturber to be, but this is a good place to start.