Logging changes.
I've done a bit of work here to (I think) improve how things get logged.

I've added a LoggingOptions type that holds the setting here.  The idea
is that we specify the equivalent of 'sample_info.txt' and 'sample.txt'
in a LoggingOptions value that is passed to 'nestedSampling'; if those
are missing, then we don't output that kind of progress at all.  There's
also a flag for logging progress to stdout.

Then I export a 'defaultLogging' value that has sane defaults.  I've
changed the default filenames to hopefully be a little more descriptive.
We may want to just default to not creating files here and let users
overwrite that if desired.

I've also gotten rid of the 'generating from prior..' message because it
doesn't seem too useful.  We could keep it in there, but would
then have to pass logging options to 'intiialize' and it doesn't seem
worth it for that message.